refactor: feeds types 수정
