+++ src/api/feeds/index.ts
@@ -1,5 +1,5 @@
 import { axiosHandler } from "@/utils/axios.utils";
-import { GetFeedsTypes } from "@/types/feeds/feedsRequestTypes";
+import { GetFeedsParamsTypes } from "@/types/feeds/feedsRequestTypes";
 
 /**
  * feeds apis
@@ -7,7 +7,7 @@ import { GetFeedsTypes } from "@/types/feeds/feedsRequestTypes";
  */
 
 export const feedsApi = Object.freeze({
-	async getFeedsRequest(endPoint: string, params: GetFeedsTypes) {
+	async getFeedsRequest(endPoint: string, params: GetFeedsParamsTypes) {
 		try {
 			const response = await axiosHandler.get(endPoint, { params });
 			return response;
@@ -15,7 +15,7 @@ export const feedsApi = Object.freeze({
 			return err;
 		}
 	},
-	async postFeedRequest(endPoint: string, params: GetFeedsTypes) {
+	async postFeedRequest(endPoint: string, params: GetFeedsParamsTypes) {
 		try {
 			const response = await axiosHandler.post(
 				endPoint,
