+++ src/types/feeds/feedsResponseTypes.ts
@@ -15,6 +15,12 @@ export interface UserDailyNutrientTypes {
 }
 
 export interface GetFeedsResponseTypes {
+	prev_page: boolean;
+	next_page: boolean;
+	feeds: GetFeedsTypes[];
+}
+
+export interface GetFeedsTypes {
 	feed_id: number;
 	user_id: number;
 	image_url: string | null;
@@ -34,9 +40,4 @@ export interface GetFeedsResponseTypes {
 	goal: GoalType;
 	my_like: boolean;
 	is_mine: boolean;
-}
-
-export interface GetPagesTypes {
-	prev_page: boolean;
-	next_page: boolean;
 }
